+++ changelog.md
@@ -1,7 +1,7 @@
 # Changelog for NovaSheets
 - [1.0.0](#100)
   - [1.0.0-pre1](#100-pre1) • [1.0.0-pre2](#100-pre2) • [1.0.0-pre3](#100-pre3) • [1.0.0-pre4](#100-pre4)
-  • [1.0.0-rc1](#100-rc1) • [1.0.0-rc2](#100-rc2) • [1.0.0-rc3](#100-rc3) • [1.0.0-rc4](#100-rc4) • [1.0.0-rc5](#100-rc5)
+  • [1.0.0-rc1](#100-rc1) • [1.0.0-rc2](#100-rc2) • [1.0.0-rc3](#100-rc3) • [1.0.0-rc4](#100-rc4) • [1.0.0-rc5](#100-rc5) • [1.0.0-rc6](#100-rc6)
 - [0.6.0](#060) • [0.6.1](#061) • [0.6.2](#062) • [0.6.3](#063) • [0.6.4](#064) • [0.6.5](#065) • [0.6.6](#066) • [0.6.7](#067)
 - [0.5.0](#050) • [0.5.1](#051) • [0.5.2](#052)
 - [0.4.0](#040) • [0.4.1](#041) • [0.4.2](#042) • [0.4.3](#043) • [0.4.4](#044) • [0.4.5](#045) • [0.4.6](#046) • [0.4.7](#047)
@@ -11,6 +11,21 @@
 
 ## 1.0.0
 *Upcoming*
+
+### 1.0.0-rc6
+*2021-06-12*
+- **Additions**
+  - [CLI] Added `[<command>]` option to `novasheets --help` to display help for a given command.
+- **Changes**
+  - [Syntax] Changed selectors to no longer distribute comma-separated segments over children.
+- **Removals**
+  - [Syntax] Removed parser constants `MAX_MATH_RECURSION` and `KEEP_NAN`.
+- **Fixes**
+  - [Syntax] Fixed variables containing braces/blocks breaking the parser.
+- **Internal**
+  - [Source] Replaced math parsing code with package `math-and-unit-parser`.
+  - [Source] Replaced nesting compilation code using package `balanced-match`.
+  - [Source] Replaced browser building code with package `webpack`.
 
 ### 1.0.0-rc5
 *2021-05-29*